requirements.txt
apache-airflow
kafka-python
pyspark
pymongo
textblob
gradio
requests
requests==2.31.0
python-dotenv==1.0.1

